# swagger/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Official-Receipt service API endpoints",
    "title": "Official-Receipt API.",
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "paths": {
    "/account/{id}": {
      "get": {
        "summary": "This will get the account with the given ID",
        "operationId": "Account",
        "parameters": [
          {
            "type": "string",
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Account"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GenericRes"
          },
          "400": {
            "$ref": "#/responses/JSONErrors"
          },
          "500": {
            "$ref": "#/responses/JSONErrors"
          }
        }
      }
    },
    "/orgs/{org}/comments": {
      "get": {
        "summary": "This will get all the comments in an organization",
        "operationId": "CommentResponse",
        "parameters": [
          {
            "type": "string",
            "name": "org",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/CommentResponse"
          },
          "400": {
            "$ref": "#/responses/JSONErrors"
          },
          "500": {
            "$ref": "#/responses/JSONErrors"
          }
        }
      },
      "post": {
        "description": "Add comment to an organization",
        "tags": [
          "org"
        ],
        "operationId": "CommentReq",
        "parameters": [
          {
            "name": "org",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "CommentReq",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/CommentReq"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GenericRes"
          },
          "400": {
            "$ref": "#/responses/JSONErrors"
          },
          "500": {
            "$ref": "#/responses/JSONErrors"
          }
        }
      },
      "delete": {
        "summary": "This will delete all the comments in an organization",
        "operationId": "Comment",
        "parameters": [
          {
            "type": "string",
            "name": "org",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/GenericRes"
          },
          "400": {
            "$ref": "#/responses/JSONErrors"
          },
          "500": {
            "$ref": "#/responses/JSONErrors"
          }
        }
      }
    },
    "/orgs/{org}/members": {
      "get": {
        "summary": "This will get all the members in an organization",
        "operationId": "AccountResponse",
        "parameters": [
          {
            "type": "string",
            "name": "org",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/AccountResponse"
          },
          "400": {
            "$ref": "#/responses/JSONErrors"
          },
          "500": {
            "$ref": "#/responses/JSONErrors"
          }
        }
      }
    }
  },
  "definitions": {
    "Account": {
      "type": "object",
      "properties": {
        "AvatarURL": {
          "type": "string"
        },
        "FollowedNum": {
          "type": "integer",
          "format": "int32"
        },
        "FollowersNum": {
          "type": "integer",
          "format": "int32"
        },
        "OrganizationID": {
          "type": "string"
        },
        "Username": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-ta.git/internal/model"
    },
    "AccountResponse": {
      "description": "AccountResponse comments response",
      "type": "object",
      "properties": {
        "accounts": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Account"
          },
          "x-go-name": "Accounts"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-ta.git/internal/model"
    },
    "CommentReq": {
      "description": "CommentReq comment request data structure",
      "type": "object",
      "properties": {
        "comment": {
          "type": "string",
          "x-go-name": "Comment"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-ta.git/internal/model"
    },
    "CommentResponse": {
      "description": "CommentResponse comments response",
      "type": "object",
      "properties": {
        "comments": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "x-go-name": "Comments"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-ta.git/internal/model"
    },
    "GenericResponse": {
      "description": "GenericResponse generic response",
      "type": "object",
      "properties": {
        "reason": {
          "type": "string",
          "x-go-name": "Reason"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-ta.git/internal/model"
    },
    "JSONError": {
      "description": "JSONError generic response",
      "type": "object",
      "properties": {
        "detail": {
          "type": "string",
          "x-go-name": "Detail"
        },
        "source": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          },
          "x-go-name": "Source"
        },
        "status": {
          "type": "string",
          "x-go-name": "Status"
        },
        "title": {
          "type": "string",
          "x-go-name": "Title"
        }
      },
      "x-go-package": "github.com/johnearl92/xendit-ta.git/internal/model/errors"
    }
  },
  "responses": {
    "AccountResponse": {
      "description": "AccountResWrapperv wrapper struct for AccountResponse",
      "schema": {
        "$ref": "#/definitions/AccountResponse"
      }
    },
    "CommentResponse": {
      "description": "CommentResWrapper wrapper struct for CommentRes",
      "schema": {
        "$ref": "#/definitions/CommentResponse"
      }
    },
    "GenericRes": {
      "description": "GenericRes response model",
      "schema": {
        "$ref": "#/definitions/GenericResponse"
      }
    },
    "JSONErrors": {
      "description": "JSONErrors response model",
      "schema": {
        "type": "array",
        "items": {
          "$ref": "#/definitions/JSONError"
        }
      }
    }
  }
}